Add share button on list
