+++ app/src/main/res/layout/task.xml
@@ -45,56 +45,29 @@
         app:layout_constraintTop_toTopOf="parent"
         android:paddingTop="70dp"
         android:paddingBottom="10dp"
-        android:visibility="gone"
-        >
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
+        android:visibility="gone">
 
         <TextView
             android:id="@+id/task_description"
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="Task detail here..."
             android:paddingStart="10dp"
             android:paddingEnd="10dp"
             android:textColor="@color/black"
-            android:textSize="18sp"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintRight_toLeftOf="@id/share"
-            />
+            android:textSize="18sp"/>
 
         <TextView
-            android:id="@+id/tags"
             android:textAlignment="textEnd"
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="#Tag1 #Tag2"
             android:paddingStart="20dp"
             android:paddingEnd="10dp"
             android:textColor="@color/grey"
-            android:textSize="18sp"
-            app:layout_constraintTop_toBottomOf="@id/task_description"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toLeftOf="@id/share"/>
-
-            <Button
-                android:id="@+id/share"
-                style="@style/AppTheme.Button.Share"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintLeft_toRightOf="@id/task_description"
-                app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toTopOf="@id/task_description"
-                 />
-        </androidx.constraintlayout.widget.ConstraintLayout>
+            android:textSize="18sp"/>
 
     </LinearLayout>
-
-
-
 
     <!-- Task View (brief) -->
     <LinearLayout
